presidio-service/requirements.txt
# Web Framework (Stand: Januar 2026)
fastapi==0.128.0
uvicorn[standard]==0.40.0
gunicorn==23.0.0
python-multipart==0.0.18

# Presidio (PII Detection & Anonymization)
presidio-analyzer==2.2.360
presidio-anonymizer==2.2.360
presidio-image-redactor==0.0.56

# NLP & OCR
pytesseract==0.3.13
pillow==11.1.0

# Document Processing
pypdf2==3.0.1
python-docx==1.1.2
pdfplumber==0.11.9
pdf2image==1.17.0

# Config
pydantic-settings==2.7.1
